matroska: do not lose the top bits when writing a > 32 bit value
Coverity 1139806

+++ gst/matroska/ebml-write.c
@@ -619,9 +619,9 @@ gst_ebml_write_sint (GstEbmlWrite * ebml, guint32 id, gint64 num)
   if (num >= 0) {
     unum = num;
   } else {
-    unum = 0x80 << (size - 1);
+    unum = ((guint64) 0x80) << (size - 1);
     unum += num;
-    unum |= 0x80 << (size - 1);
+    unum |= ((guint64) 0x80) << (size - 1);
   }
 
   /* write */
